refactor(mac-embedder): thread user-event sink, split AutomationHost, strip dead events
- content: document that needs_paint's Arc<AtomicBool> is a blitz ShellProvider
  Send+Sync constraint, not our own cross-thread sharing.
- mac-embedder: move impl AutomationHost for MacApp into app/automation_host.
- mac-embedder: replace the USER_EVENT_SINK global with a threaded sink
  (EventLoopEmbedder.self.sink, a sink param on clipboard helpers, and a
  captured sink + readiness flag in the CDP automation bridge).
- mac-embedder: remove dead NewWebContentScene / CreateWindow events and a
  stale #[allow(dead_code)] on NewWebview.
